Fix bug in SuffixArray (#18)
This code failed on input string `bababa`, as the number of iterations performed was ONE short of what were required to produce the correct result. Normally, I would fix this by using a `while (true)` loop and checking if `rank == n`, see example [here](https://pastebin.com/RtB6PnEu), but in this case, the size of the matrix was hardcoded to [logn][n] so just increasing logn was sufficient.